+++ test/acceptance/test.sh
@@ -100,5 +100,14 @@ expect_2='secondary.vagrant.devcommands'
   exit 1
 }
 
+## COMPLETION-DATA
+result=$(trim "$(bundle exec vagrant run completion-data | wc -w)")
+
+[[ "${result}" == '4' ]] || {
+  echo 'Completion data contains unexpected number of return values...'
+  echo "Got result: '${result}'"
+  exit 1
+}
+
 # finish
 echo 'Success!'